requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
typer>=0.12.0
rich>=13.0.0
pyyaml>=6.0
tenacity>=8.2.0
